x86: don't accidentally zero the upper 32-bits of the PAT due to shifting past the width of a uint32_t

+++ src/system/kernel/core/processor/x86/Processor.cc
@@ -120,7 +120,7 @@ PAT7 UC
   asm volatile("rdmsr" : "=a" (pat_lo), "=d" (pat_hi) : "c" (0x277));
 
   union pat p;
-  p.x = pat_lo | (pat_hi << 32ULL);
+  p.x = pat_lo | (static_cast<uint64_t>(pat_hi) << 32ULL);
   p.s.pa0 = PAT_WB;
   p.s.pa1 = PAT_WT;
   p.s.pa2 = PAT_UCMINUS;
